# Use the shaper in math