--- a/crates/typst-layout/src/math/fragment.rs
+++ b/crates/typst-layout/src/math/fragment.rs
@@ -437,6 +437,8 @@ impl GlyphFragment {
     ///
     /// The resulting frame may not have the exact desired width or height.
     pub fn stretch(&mut self, ctx: &mut MathContext, target: Abs, axis: Axis) {
+        self.reset_glyph();
+
         // If the base glyph is good enough, use it.
         let mut advance = self.size.get(axis);
         if axis == Axis::X && !self.extended_shape {