… crates/typst-layout/src/math/lr.rs
… crates/typst-layout/src/math/mat.rs
… crates/typst-layout/src/math/mod.rs
… crates/typst-layout/src/math/root.rs
… crates/typst-layout/src/math/shared.rs
… crates/typst-layout/src/math/stretch.rs
… crates/typst-layout/src/math/text.rs
… crates/typst-layout/src/math/underover.rs
… crates/typst-library/src/text/font/mod.rs
… crates/typst-library/src/text/item.rs
… crates/typst-library/src/text/mod.rs
… crates/typst-pdf/src/text.rs
… crates/typst-render/src/text.rs
… crates/typst-svg/src/text.rs
… tests/ref/math-accent-dotless-greedy.png
… tests/suite/math/accent.typ