## Changes

The main change this PR makes is to use the shaper when laying out glyphs in math, replacing the handwritten code we had for applying `ssty`, `dtls`, etc. This also sets us up to support grapheme clusters (e.g. variation sequences) in math as well as RTL math layout in the future. To accomplish this switching a glyph to its `ssty`, `dtls`, etc. variant has been moved to the `StyleChain`.

The `GlyphFragment` struct in math is no longer limited to a single Unicode character. The `try_new` method can now take any string (intended to just be a single grapheme cluster), but will return `None` if the result from shaping isn't a single glyph.

Finally, I also culled the `VariantFragment` struct from math entirely. It really only exists because glyph assemblies need to be stored somehow. But we can actually store them in the `TextItem` that `GlyphFragment`s end up getting turned into! For horizontal assemblies there is no issue, but for vertical ones I needed to add support in the exporters for the `y_offset` and `y_advance` (it is kind of like support for very rudimentary vertical text layout has been added).
- For `typst-pdf` things were simple enough, I just modified the `impl krilla::text::Glyph for PdfGlyph` in `typst-pdf/src/text.rs` and `krilla` produces a correct looking result.
- For `typst-svg`, wherever I saw an `x` or `offset` in `typst-svg/src/text.rs` I just duplicated them to add their `y` counterparts, and correct output is produced.
- For `typst-render` I originally did the same as I did for `typst-svg`, but things were going in the wrong direction vertically. So I negated all the `y` counterparts in `typst-render/src/text.rs` and things were working. I have no idea why the y-axis is seemingly flipped in `typst-render`.

For all of these I have probably made some mistakes, so I would appreciate if others could check my changes here thoroughly!

The change in `typst-pdf` has caused the (hopefully) **only** user facing change in this PR: Text selection of glyph assemblies in math will now only have one copy of the glyph instead of a bunch of them. See below for two example PDFs and what the text selection produces with `zathura`.
<details>

<summary>Glyph assembly text selection</summary>

```typ
--- playground ---
#set page(width: auto)
$ f : A stretch(->, size: #5em) B quad A stretch(\[, size: #5em) B $
```

**Old:** [old.pdf](https://github.com/user-attachments/files/20451171/old.pdf)
```
𝑓: 𝐴→→→→→→→→→→→→→→→𝐵
𝐴
[
[[[
[
𝐵
```

**New:** [new.pdf](https://github.com/user-attachments/files/20451173/new.pdf)
```
𝑓: 𝐴→
𝐵
𝐴
[
𝐵
```


</details>

## Tests

- ~~The eight tests that have been updated are identical to the ones in #6320. This change occurred as the `ssty` OpenType feature tag is now set automatically when the `MathSize` is `Script` or `Sscript`. (I think we will probably go through with #6320 anyways, so this should be fine.)~~
  #6320 has been merged

- Having the `dtls` OpenType feature tag now live on the `StyleChain` has lead to https://github.com/MicrosoftDocs/typography-issues/issues/1191 being introduced. To track this a new test has been added:
  <details>

  <summary>math-accent-dotless-greedy</summary>

  ![math-accent-dotless-greedy](https://github.com/user-attachments/assets/39f03c5e-c61e-4efd-8320-dd0fd6bf3285)

  ```typ
  --- math-accent-dotless-greedy ---
  // Currently the dotless style propogates to everything in the accent's base,
  // even though it shouldn't.
  $ arrow(P_(c, i dot j) P_(1, i) j) \
    arrow(P_(c, i dot j) P_(1, i) j, dotless: #false) $
  ```

  </details>

- ~~There is only a single pesky test failing, math-accent-sym-call. ~~I'm investigating what's going on here...~~ I've tracked down why its failing: the glyph `ä`'s width has changed from `5.0pt` to `5.11pt` somehow. Now to figure out the cause...~~
  Fixed! No tests are failing now. (I made a mistake when calculating the accent_attach.)

## Performance

I imagine that now running everything in math through the shaper will have some performance overhead. ~~I'll try look into this soon.~~
I ran some benchmarks between `main`, this PR, and this PR but with `#[comemo::memoize]` on `GlyphFragment::try_new`, on [masterproef](https://github.com/Dherse/masterproef) and [notes](https://github.com/jthulhu/notes) with `hyperfine` (with `--warmup 3`). Here are the results:
| Command | Mean [s] | Min [s] | Max [s] | Relative |
|:---|---:|---:|---:|---:|
| `./typst-main c --root masterproef masterproef/masterproef/main.typ` | 3.823 ± 0.043 | 3.781 | 3.901 | 1.00 |
| `./typst-pr c --root masterproef masterproef/masterproef/main.typ` | 3.839 ± 0.028 | 3.802 | 3.886 | 1.00 ± 0.01 |
| `./typst-pr-opt c --root masterproef masterproef/masterproef/main.typ` | 3.875 ± 0.038 | 3.802 | 3.916 | 1.01 ± 0.02 |

| Command | Mean [s] | Min [s] | Max [s] | Relative |
|:---|---:|---:|---:|---:|
| `./typst-main c notes/grothendieck-fibrations.typ` | 14.306 ± 0.063 | 14.218 | 14.376 | 1.01 ± 0.01 |
| `./typst-pr c notes/grothendieck-fibrations.typ` | 14.404 ± 0.076 | 14.327 | 14.544 | 1.01 ± 0.01 |
| `./typst-pr-opt c notes/grothendieck-fibrations.typ` | 14.196 ± 0.044 | 14.086 | 14.240 | 1.00 |