@LaurenzV 2025-05-27
> @LaurenzV would you happen to know if this is expected behaviour in rustybuzz?

Hmm, I'm not sure, but at least the `SCRIPT_MATH` constant isn't used anywhere. @behdad Do you know if it does anything in harfbuzz?