@mkorje 2025-05-28
I opened https://github.com/harfbuzz/rustybuzz/issues/162.