@laurmaedje 2025-06-09
I agree with @wrzian, the PR looks quite excellent! I don't think there's too much to say either way, but I think it would still be good to rebase before I leave my review.